+++ plotnine/geoms/geom_text.py
@@ -1,4 +1,5 @@
 from contextlib import suppress
+from warnings import warn
 
 from matplotlib.text import Text
 
@@ -9,6 +10,7 @@ except ImportError:
 else:
     HAS_ADJUST_TEXT = True
 
+from ..exceptions import PlotnineWarning
 from ..utils import to_rgba, order_as_data_mapping
 from ..doctools import document
 from ..positions import position_nudge
@@ -46,7 +48,11 @@ class geom_text(geom):
         Parameters to :class:`adjustText.adjust_text` will repel
         overlapping texts. This parameter takes priority of over
         ``nudge_x`` and ``nudge_y``.
-        See https://github.com/Phlya/adjustText/wiki .
+
+        ``adjust_text`` does not work well when it is used in the
+        first layer of the plot, or if it is the only layer.
+        For more see the documentation at
+        https://github.com/Phlya/adjustText/wiki .
     format_string : str (default: None)
         If not :py:`None`, then the text if formatted with this
         string using :meth:`str.format`
@@ -58,6 +64,7 @@ class geom_text(geom):
 
     See Also
     --------
+    plotnine.geoms.geom_label
     matplotlib.text.Text
     matplotlib.patheffects
 
@@ -179,6 +186,8 @@ class geom_text(geom):
         else:
             bbox = {}
 
+        texts = []
+
         # For labels add a bbox
         for i in range(len(data)):
             kw = df.iloc[i].to_dict()
@@ -187,11 +196,18 @@ class geom_text(geom):
                 kw['bbox']['edgecolor'] = params['boxcolor'] or kw['color']
                 kw['bbox']['facecolor'] = kw.pop('facecolor')
             text_elem = ax.text(**kw)
+            texts.append(text_elem)
             if params['path_effects']:
                 text_elem.set_path_effects(params['path_effects'])
 
         if params['adjust_text']:
-            adjust_text(list(ax.texts), ax=ax, **params['adjust_text'])
+            if params['zorder'] == 1:
+                warn(
+                    "For better results with adjust_text, it should "
+                    "not be the first layer or the only layer.",
+                    PlotnineWarning
+                )
+            adjust_text(texts, ax=ax, **params['adjust_text'])
 
     @staticmethod
     def draw_legend(data, da, lyr):